Allow specific callback for go to definition
This allows users to set g:ale_go_to_definition_callback and then
call that instead of opening in a buffer or tab.

The motivation is that users may have specific logic on what needs to
happen when a definition is found, like looking to check whether the
specific file is already in a tab, then go to tab instead of opening a
new one.

This is backwards compatible as the default maintains current behavior.